+++ QueProject/QueProject/Configs/Configurations.swift
@@ -7,29 +7,7 @@
 
 import Foundation
 
-@propertyWrapper
-struct BundleProperty<T> {
-    private let key: String
-    
-    init(key: String) {
-        self.key = key
-    }
-    
-    var wrappedValue: T {
-        guard let value = Bundle.main.infoDictionary?[key] as? T else {
-            fatalError("Key \(key) is not set in Info.plist")
-        }
-        return value
-    }
-}
-
-extension Bundle {
+struct APIConfig {
     @BundleProperty(key: "BaseURL")
     static var baseURL: String
-}
-
-struct APIConfig {
-    static var baseURL: String {
-        Bundle.baseURL
-    }
 }
